Revert "#1 petita prova (#1)"
This reverts commit 882147ce40c281ae034bc3949a381051a346f877.

+++ 2.Shared/Models/Money.cs
@@ -2,7 +2,7 @@
 {
     public record Money
     {
-        public double Amount { get; init; } 
+        public decimal Amount { get; init; } 
         
         public Currencies Currency { get; init; }
